webserver/c4Lookup/requirements.txt
attrs==19.1.0
autopep8==1.4.4
backports.csv==1.0.7
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.2
django-cors-headers==3.0.2
django-crispy-forms==1.7.2
django-environ==0.4.5
django-filter==2.1.0
django-filters==0.2.1
django-import-export==1.2.0
django-rest-auth==0.9.5
djangorestframework==3.9.4
drf-yasg==1.16.0
et-xmlfile==1.0.1
idna==2.8
inflection==0.3.1
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
jsonschema==3.0.1
MarkupSafe==1.1.1
mysqlclient==1.4.2.post1
odfpy==1.4.0
openpyxl==2.6.2
pycodestyle==2.5.0
pyrsistent==0.15.2
pytz==2019.1
PyYAML==5.1
requests==2.22.0
rest-framework-latex==0.1.0
ruamel.yaml==0.15.97
sentry-sdk==0.9.1
six==1.12.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
tablib==0.13.0
uritemplate==3.0.0
urllib3==1.25.3
waitress==1.3.0
whitenoise==4.1.2
xlrd==1.2.0
xlwt==1.3.0
